+++ .gitignore
@@ -4,6 +4,7 @@ push_swap_debug
 
 # Checker
 checker_Mac
+pyviz.py
 
 # Prerequisites
 *.d
