+++ includes/opers.h
@@ -18,6 +18,7 @@
 ** < handler.c > */
 
 t_res		oper(t_engine *engine, t_op op);
+t_res		inst(t_engine *engine, t_flag what, t_inst inst);
 /*
 ** < handler_utils.c > */
 
@@ -28,7 +29,7 @@ void		set_deques_from_to(t_engine *engine, t_deque *deqs[2],
 				t_flag from);
 const char	*get_op_name(t_op op);
 /*
-** < operations.c > */
+** < opers.c > */
 
 t_res		oper_swap(t_engine *engine, t_flag what);
 t_res		oper_push(t_engine *engine, t_flag to);
